# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create an account and sign in",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Sign in with email and password",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Exchange a refresh token for a new access token",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Revoke a refresh token",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/venues/{id}/join": {
      "post": {
        "operationId": "PresenceController_join",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationFixDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Join the venue chat; requires a precise fix within the join radius",
        "tags": [
          "venues"
        ]
      }
    },
    "/venues/{id}/leave": {
      "post": {
        "operationId": "PresenceController_leave",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Leave the venue chat",
        "tags": [
          "venues"
        ]
      }
    },
    "/users/{id}/block": {
      "post": {
        "operationId": "BlocksController_block",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Block a user (they are not told)",
        "tags": [
          "blocks"
        ]
      },
      "delete": {
        "operationId": "BlocksController_unblock",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Unblock a user",
        "tags": [
          "blocks"
        ]
      }
    },
    "/me/blocks": {
      "get": {
        "operationId": "BlocksController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Users I have blocked",
        "tags": [
          "blocks"
        ]
      }
    },
    "/venues/{id}/messages": {
      "post": {
        "operationId": "MessagesController_send",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Send a message; it is moderated before anyone sees it",
        "tags": [
          "messages"
        ]
      },
      "get": {
        "operationId": "MessagesController_history",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "afterId",
            "required": false,
            "in": "query",
            "description": "Return messages newer than this message id (gap fill after reconnect)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Recent visible messages (oldest first), or everything after a message id",
        "tags": [
          "messages"
        ]
      }
    },
    "/messages/{id}/reports": {
      "post": {
        "operationId": "ReportsController_create",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReportDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Report a message for moderator review",
        "tags": [
          "reports"
        ]
      }
    },
    "/admin/reports": {
      "get": {
        "operationId": "AdminController_reports",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "default": "OPEN",
              "type": "string",
              "enum": [
                "OPEN",
                "REVIEWED",
                "ACTIONED",
                "DISMISSED"
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reports queue",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/incidents": {
      "get": {
        "operationId": "AdminController_incidents",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "default": "OPEN",
              "type": "string",
              "enum": [
                "OPEN",
                "RESOLVED"
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Automatic incidents (severe content, strike thresholds, report thresholds)",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/messages/{id}": {
      "get": {
        "operationId": "AdminController_message",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Full message record including original text and reports",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/reports/{id}/resolve": {
      "post": {
        "operationId": "AdminController_resolve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveReportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Resolve a report: dismiss, hide, mute or suspend",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/incidents/{id}/resolve": {
      "post": {
        "operationId": "AdminController_resolveIncident",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark an incident as handled",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/users/{id}/suspend": {
      "post": {
        "operationId": "AdminController_suspend",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuspendUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Suspend an account",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/users/{id}/unsuspend": {
      "post": {
        "operationId": "AdminController_unsuspend",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lift a suspension",
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Larea API",
    "description": "Location-based chat backend. Realtime events are documented in docs/realtime-protocol.md.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "anna@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 10,
            "maxLength": 128
          },
          "displayName": {
            "type": "string",
            "example": "anna_k",
            "description": "Display name must be 3-24 characters: letters, numbers, spaces, underscores or dots."
          },
          "deviceLabel": {
            "type": "string",
            "example": "iPhone 15"
          }
        },
        "required": [
          "email",
          "password",
          "displayName"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "anna@example.com"
          },
          "password": {
            "type": "string"
          },
          "deviceLabel": {
            "type": "string",
            "example": "Pixel 8"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "LogoutDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
...
      "LocationFixDto": {
        "type": "object",
        "properties": {
          "lat": {
            "type": "number",
            "example": 52.5219
          },
          "lng": {
            "type": "number",
            "example": 13.4132
          },
          "accuracy": {
            "type": "number",
            "description": "Horizontal accuracy radius in metres (68% confidence)",
            "example": 12
          },
          "mocked": {
            "type": "boolean",
            "description": "True when the OS reports a mock/simulated location"
          }
        },
        "required": [
          "lat",
          "lng",
          "accuracy"
        ]
      },
      "SendMessageDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 500,
            "example": "Anyone want to get food?"
          },
          "clientKey": {
            "type": "string",
            "description": "Client-generated idempotency key (e.g. a UUID); resending with the same key returns the original result"
          }
        },
        "required": [
          "text",
          "clientKey"
        ]
      },
      "CreateReportDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "enum": [
              "HARASSMENT",
              "THREAT",
              "HATE",
              "SEXUAL",
              "SPAM",
              "SCAM",
              "PERSONAL_INFO",
              "OTHER"
            ]
          },
          "details": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": [
          "reason"
        ]
      },
      "ResolveReportDto": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "DISMISS",
              "HIDE_MESSAGE",
              "MUTE",
              "SUSPEND"
            ]
          },
          "durationHours": {
            "type": "number",
            "description": "Mute length in hours (MUTE only)",
            "default": 24
          },
          "note": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": [
          "action"
        ]
      },
      "SuspendUserDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": [
          "reason"
        ]
      }
    }
  }
}